requirements.txt
flask==2.0.1
crawl4ai==0.5.0.post8
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule==1.2.0
python-dotenv>=1.0.0
pandas==2.0.3
apscheduler==3.10.1
tqdm==4.62.3
PyPDF2>=3.0.0
google-generativeai>=0.3.2 